requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
python-dotenv==1.0.0
marshmallow==3.20.2
flask-cors==4.0.0
cryptography==41.0.7
SQLAlchemy==2.0.23
PyJWT==2.8.0
Flask-Bcrypt==1.0.1
Werkzeug==3.0.1
ipdb==0.13.13
